// src/app/(root)/(admin)/administrator/(cms)/cms/types/site-settings.ts
import { ObjectId } from 'mongodb'

export interface SiteSettings {
	_id: ObjectId | string
	siteKeywords: string[]
	semanticCore: string[]
	metaDescription: string
	siteTitle: string
	updatedAt: string
}

export interface FormData {
	siteTitle: string
	metaDescription: string
	siteKeywords: string
	semanticCore: string
}

export interface CurrentSettingsProps {
	settings: SiteSettings
}

export interface FormButtonsProps {
	saving: boolean
	disabled?: boolean
}

export interface SEOFormProps {
	formData: FormData
	setFormData: (data: FormData) => void
	settings: SiteSettings | null
	saving: boolean
	handleSave: (e: React.SyntheticEvent) => void
	reloading?: boolean
}

export interface FormFieldProps {
	label: string
	value: string
	onChange: (value: string) => void
	type: 'text' | 'textarea'
	placeholder: string
	hint: string
	rows?: number
	showCommaHint?: boolean
	disabled?: boolean
}